Move sorters in their own package
