+++ pkg/appgw/backendaddresspools.go
@@ -7,6 +7,7 @@ package appgw
 
 import (
 	"fmt"
+	"github.com/Azure/application-gateway-kubernetes-ingress/pkg/sorter"
 	"sort"
 
 	n "github.com/Azure/azure-sdk-for-go/services/network/mgmt/2018-12-01/network"
@@ -122,7 +123,7 @@ func getBackendAddressMapKeys(m *map[n.ApplicationGatewayBackendAddress]interfac
 	for addr := range *m {
 		addresses = append(addresses, addr)
 	}
-	sort.Sort(byIPFQDN(addresses))
+	sort.Sort(sorter.ByIPFQDN(addresses))
 	return &addresses
 }
